검색 기능, 유저 검색기능 테스트완료, 이미지 출력 및 프로필 업데이트 기능 적용 해야할 것: bio 수정기능, subtag 검색시 본인의 subtag list 나오게, 더보기 기능

+++ src/main/java/com/vreads/backend/mapper/VreadsMapper.java
@@ -16,6 +16,12 @@ public interface VreadsMapper {
 	
 	
 	//특정 유저 Vreads 불러오기
+	/*
+	 * <!-- serchType -->
+	<!-- 1 = 제목검색 -->
+	<!-- 2 = 제목과 타이틀 검색 -->
+	<!-- 3 = 서브태그 검색 -->
+	 * */
 	List<VreadsVO> selectUserVreads(
 			@Param("userIdx") int userIdx, 
 			@Param("keyword") String keyword, 
@@ -25,7 +31,7 @@ public interface VreadsMapper {
 			@Param("setPageListLimit") int setPageListLimit);
 
 	//특정 Vreads 하나만 불러오기
-	VreadsVO selectVreadDetail(@Param("vreadsIdx") int vreadsIdx);
+	VreadsVO selectVreadDetail(@Param("vreadsIdx") Long vreadsIdx);
 
 	//Vreads 조건 없이 불러오기 
 	List<VreadsVO> selectAllVreads(
@@ -34,6 +40,12 @@ public interface VreadsMapper {
 			@Param("setPageListLimit") int setPageListLimit);
 
 	//search type 과 keyword 에 맞춰서 Vreads 불러오기
+	/*
+	 * 	<!-- serchType -->
+	<!-- 1 = 제목검색 -->
+	<!-- 2 = 제목과 타이틀 검색 -->
+	<!-- 3 = 서브태그 검색 -->
+	 * */
 	List<VreadsVO> selectSearchVreads(
 			@Param("keyword") String keyword, 
 			@Param("searchType") int searchType,
